% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce_migr}
\alias{coalesce_migr}
\title{Performs coalescence and output multidimensional site frequency spectra (SFS)}
\usage{
coalesce_migr(x, nRep = 1, SAA = 1, execute = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{an initialized reference table produced by \code{\link{generate_priors}}}

\item{nRep}{(int) number of repetitions for each parameter combination (only for debug mode, otherwise prefer the default value)}

\item{execute}{(logical) if TRUE, execute the coalescence}

\item{verbose}{(logical) if TRUE, print all messages}
}
\value{
A reference table, of class \code{referenceTable}, with all simulated multidimensional site frequency spectra.
}

