% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_old.R
\name{coalesce_old}
\alias{coalesce_old}
\title{Performs multicoalescence SIMPLE PARTITION WITHOUT MIGRATION, given a set of parameter priors}
\usage{
coalesce_old(x, method = "NULL", nRep = 1, SAA = 1, execute = TRUE,
  phyclust = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{an initialized reference table (class \emph{referenceTable}).}

\item{nRep}{(optional) number of repetitions to approximate the sweep (default 100)}
}
\value{
A reference table with all simulated site frequency spectra.
}
\description{
NA
}
\details{
Past resizings are always set in reference to  \emph{No}.
}

