% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Compute a confusion matrix}
\usage{
confusion_matrix(cvd)
}
\arguments{
\item{cvd}{a matrix of strings with two columns: (1) the true model; (2) the predicted model}
}
\value{
A confusion matrix.
}
\keyword{internal}

