% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Dimension reduction for multidimensional SFSs}
\usage{
dim_reduction(x, removeCollinearCols = TRUE, tolLDA = 1e-04, minSDlda = 0,
  PLS_normalize = TRUE, PLS_ncomp = NULL, doPLSrecRate = TRUE,
  n_dist_bstp_per_model = 2000, relativeSFS = FALSE,
  nonvarTolerance = 1e-04, abcTolerance = 0.01,
  QR_multicollinearity = TRUE, plot_PLScv = FALSE)
}
\arguments{
\item{x}{(referenceTable) a reference table with \code{SFS} elements}

\item{removeCollinearCols}{(logical) whether to remove collinear columns across the multiSFSs (TRUE is recommended for LDA analysis)}

\item{PLS_normalize}{(logical) whether to normalize the multiSFSs prior to PLS analysis (TRUE is recommended)}

\item{PLS_ncomp}{(integer) if NULL, automatically select the fittest number of PLS components, else an integer to manually set the number of components to retain}

\item{doPLSrecRate}{(logical) whether to regress (in addition to the sweep ages) onto the recombination parameter in the PLS analyses (since the recombination rate is considered a nuisance parameter, it can be sometimes useful to discard it from the PLS analysis)}

\item{n_dist_bstp_per_model}{(integer) number of jacknifes to perform per model to compute the null distribution of distances for later goodness-of-fit tests}

\item{relativeSFS}{(logical) if TRUE, will use relative SFS, else will use absolute SFS (with absolute SNP counts per frequency class) (FALSE is recommended)}

\item{nonvarTolerance}{(numeric) variance threshold across the multiSFSs for a given frequency class, if the variance is below this threshold, the frequency class will be discarded for LDA analysis}

\item{abcTolerance}{(numeric < 1) ABC tolerance ratio for the distance-based goodness-of-fit test (\emph{cf. n_dist_btsp_per_model} and see \code{\link{abc}})}

\item{plot_PLScv}{(logical) if TRUE, will plot the cross-validation graphs of the PLS}
}
\value{
An object of class \code{referenceTable} with filled-in \code{DIMREDUC} element.
}
\description{
Performs a series of dimensional reductions: LDA (for later model selection) and model-specific PLSs (for later parameter estimation).
}

