% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction_rf.R
\name{dim_reduction_rf}
\alias{dim_reduction_rf}
\title{LDA & RF supervised machine learning of the selection signals}
\usage{
dim_reduction_rf(x, removeCollinearCols = TRUE, compress_SFS = TRUE,
  LDA_xPC = 0.1, LDA_sizePerModel = NULL, dating_method = "rf",
  PLS_normalize = TRUE, PLS_ncomp = NULL, PLS_ncomp_method = "elbow",
  PLS_maxncomp = 100, PLS_plotCV = TRUE, RF_ntree = 500, RF_pcomp = 1,
  ...)
}
\arguments{
\item{x}{a \code{referenceTable} object with non-empty \code{PRIORS} and \code{SFS} slots}

\item{removeCollinearCols}{(logical) whether to remove collinear columns from the site frequency spectra prior to LDA & PLS dimension reduction (TRUE is recommended)}

\item{LDA_xPC}{(float between 0 and 1) number of principal components to retain as a fraction of the original number of cells in the SFS; LDA will be trained on the PCA-projected simulations}

\item{LDA_sizePerModel}{(integer) number of simulations per model to consider for training the LDA; \emph{NULL} (by default) forces to use all data}

\item{PLS_normalize}{(logical) whether to normalize the spectra prior to PLS analysis (TRUE is recommended)}

\item{PLS_ncomp}{(integer) if NULL, automatically select the fittest number of PLS components, else an integer to manually set the number of retained components}

\item{PLS_ncomp_method}{(string) the method to automatically select the best number of features: either "\emph{elbow}", "\emph{onesigma}" or "\emph{randomization}" (cf. \code{selectNcomp} in package \code{pls})}

\item{PLS_maxncomp}{(integer) if automatic selection of PLS features, the maximum number of PLS components to consider before downsampling to the fittest set}

\item{RF_ntree}{(integer) number of trees to grow for the random forest (has a quadratic effect on computation time)}

\item{RF_pcomp}{(numeric, between 0 and 1) proportion of features to retain for the random forest modelling}

\item{...}{other arguments passed to \code{randomForest}}

\item{plot_PLScv}{(logical) if TRUE, plots the PLS cross-validation graphs}
}
\value{
An object of class \code{referenceTable} with filled-in \code{DIMREDUC} slot.
}
\description{
Performs supervised learning algorithms (LDA and random forests) on the summary statistics (joint multidimensional allele frequency spectra) prior to genome scan. This function is to be run after \code{\link{coalesce}}.
}
\examples{
Please refer to the vignette.
}
\seealso{
\code{\link{coalesce}}, \code{\link{plsr}}, \code{\link{lda}}
}

