% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{explain}
\alias{explain}
\title{explain a \code{referenceTable} object}
\usage{
explain(X)
}
\arguments{
\item{X}{a \code{referenceTable} object}
}
\value{
A summary of the content of the reference table.
}
\seealso{
\code{\link{generate_priors}}
}

