% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_priors.R
\name{generate_priors}
\alias{generate_priors}
\title{Initialize the reference table}
\usage{
generate_priors(msDemography, No, fold, windowSize = 20000, nSimul = 10000,
  restrictDemesTo = NULL, sweepAge = NULL, recRate = list("rlogunif",
  1e-12, 0.001))
}
\arguments{
\item{msDemography}{(string) the neutral demographic history of the metapopulation, formatted according to Hudson's \code{\link{ms}} conventions, where \emph{theta} equals \eqn{4*No*mutRate}}

\item{No}{(int) effective size of the reference population (i.e. number of diploid individuals)}

\item{windowSize}{(int) window size for the genome scan (number of base pairs)}

\item{nSimul}{(int) number of simulations to perform for each model}

\item{restrictDemesTo}{(array of int) by default, McSwan will detect positive sweeps across the recent history of \bold{every} deme specified in the \code{ms} command; however, if you want to restrict the analysis to specific demes, provide their \bold{indices} here in a vector (NB. the deme index corresponds to its position under the \code{-I} option of the \code{ms} command, note that indexation is 1-based)}

\item{sweepAge}{(special list) prior distribution for the sweep ages (scaled in generations before present) (if \code{NULL} and the number of demes is superior to 2, the prior distribution will be automatically set, otherwise it is mandatory to specify the distribution manually, see \link{Details})}

\item{recRate}{(special list) prior distribution for the recombination rates (see \link{Details})}
}
\value{
An object of class \emph{referenceTable} with initialized prior fields.
}
\description{
Semi-automatic initialization of the Bayesian prior distributions of parameters across all possible demographic models: neutral and with local positive adaptation in each specific deme.
}
\details{
Prior distributions must be specified using the following syntax: \code{list("P", arg1, arg2)} with \emph{P} the name of the distribution function (e.g. \code{\link{runif}} for the uniform distribution, \code{\link{rlogunif}} for the log-uniform; please make sure you have quoted the function name and removed the argument brackets); \emph{arg1} and \emph{arg2} respectively the first and second arguments of the function (e.g. for \code{runif} will be the lower and upper limits of the distribution). Note that if the upper and lower limits of the distribution are distant by more than 3 orders of magnitude in \emph{log10}, we advise to use the \code{rlogunif} distribution.

If you manually provide the prior distributions for the \bold{sweep ages}, you have two options: \itemize{
\item either specify uniquely \code{list("P", arg1, arg2)} which will be the distribution set \bold{across} the demes,
\item or specify a list of distribution-lists, each distribution-list corresponding to the sweep age distribution for a specific deme (indexed as they appear in the \code{ms} command); for instance, for 2 demes, one would specify: \code{list(list("rlogunif", T_1, TT_1), list("runif", T_2, TT_2))}. Please note that you must specify distributions for every deme, even if you have restricted the analysis to specific demes \emph{via} the \code{restrictDemesTo} option.
}
}
\examples{
mut <- 2.5e-8
No <- 2053
reftb <- generate_priors(msDemography = paste("80 1 -t",4*No*mut,
                                             "-I 2 40 40",
                                             "-en 0 2",29153/No,
                                             "-en",70/(4*No),2,11869/No,
                                             "-ej",2000/(4*No),"1 2",
                                             "-en",2683/(4*No),2,7083/No),
                        No = No, windowSize = 20000, nSimul = 1e3)
}
\references{
Hudson, R.R. (2007) \code{ms} - a program for generating samples under neutral models \url{http://home.uchicago.edu/rhudson1/source/mksamples/msdir/msdoc.pdf}
}
\seealso{
\code{\link{coalesce}}, \code{\link{get_max_sweep_ages}}
}

