% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sweep_pseudochr.R
\name{generate_pseudoobs}
\alias{generate_pseudoobs}
\title{Generates pseudo-observed genomic fragments under arbitrary demographic histories with and without positive sweeps}
\usage{
generate_pseudoobs(reftb, nSimul = 50, sweepingIsl = NULL, recRate,
  sweepAge = NULL, L = NULL, sweepPos = 0.5, Smu = 0.01, nReps = 1,
  verbose = FALSE, doSFS = TRUE, default_sweepAge_prior_func = "rlogunif",
  save_each_file = FALSE)
}
\arguments{
\item{reftb}{an initialized \emph{referenceTable} object}

\item{nSimul}{(integer) number of independent genomic fragments to simulate}

\item{recRate}{(special list) prior distribution for the recombination rates (see \link{Details})}

\item{sweepAge}{(special list) prior distribution for the sweep ages (scaled in generations before present) (if \code{NULL} and the number of demes is superior to 2, the prior distribution will be automatically set, otherwise it is mandatory to specify the distribution manually, see \link{Details})}

\item{L}{(integer) genomic fragment length (in base pairs); if \code{NULL}, the function will perform a non-sliding window validation; if not-\code{NULL}, the function will perform sliding window validation (closer to the genomic scan approach used to analyze observed dataset)}

\item{sweepPos}{(numeric) relative position of the beneficial mutation (eg. for \eqn{sweepPos=0.5}, the beneficial mutation will be placed at the center of the genomic region)}

\item{Smu}{forward mutation rate to the beneficial allele}

\item{nReps}{(integer) number of simulations per parameter combination (by default 1, but you can increase it to reduce the stochasticity of the coalescent process, the function will automatically average the coalescent-generated multiSFSs)}

\item{verbose}{(logical) whether to print all messages (TRUE) or not (FALSE)}

\item{doSFS}{(logical) whether to compute the multiSFS (TRUE) or not (FALSE)}
}
\value{
An object with special class (see below) containing in the \code{SFS} element: \itemize{
\item multidimensional SFSs if \eqn{L = NULL} (class \code{validationTable})
\item positional allele frequencies if \eqn{L != NULL} (class \code{slidingValidationTable})
}
}
\description{
Simulation based on Ewing's \emph{et al.} \code{msms}. Assumes very strong selection coefficients.
}
\details{
Prior distributions must be specified using the following syntax: \code{list("P", arg1, arg2)} with \emph{P} the name of the distribution function (e.g. \code{\link{runif}} for the uniform distribution, \code{\link{rlogunif}} for the log-uniform; please make sure you have quoted the function name and removed the argument brackets); \emph{arg1} and \emph{arg2} respectively the first and second arguments of the function (e.g. for \code{runif} will be the lower and upper limits of the distribution). Note that if the upper and lower limits of the distribution are distant by more than 3 orders of magnitude in \emph{log10}, we advise to use the \code{rlogunif} distribution.

If you manually provide the prior distributions for the \bold{sweep ages}, you have to respect the following format: \itemize{
\item specify a list of distribution-sublists, each distribution-sublist corresponding to the sweep age distribution for the specific deme(s) you provided in the \code{sweepingIsl} argument (indexed as they appear in the \code{ms} command) (if \eqn{sweepingIsl = NULL} you will need to provide the distribution-sublists for \bold{all} demes); for instance, for \eqn{sweepingIsl = c(1,2)}, one would specify: \code{list(list("rlogunif", T_1, TT_1), list("runif", T_2, TT_2))}.
}
}
\references{
Ewing et Hermisson (2010) MSMS: a coalescent simulation program including recombination, demographic structure and selection at a single locus. \emph{Bioinformatics}.
}

