% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{get_dists}
\alias{get_dists}
\title{Get the null distribution of distances to some pseudo-observed datastes}
\usage{
get_dists(x, n_dist_btsp_per_model, abcTolerance)
}
\arguments{
\item{x}{a model-specific multiSFS matrix}

\item{n_dist_btsp_per_model}{number of distances to perform per model for computation of the null distribution}
}
\value{
A vector of null distances.
}
\keyword{internal}

