% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_priors.R
\name{get_max_sweep_ages}
\alias{get_max_sweep_ages}
\title{get_max_sweep_ages}
\usage{
get_max_sweep_ages(ms, No)
}
\arguments{
\item{ms}{(string) the \code{ms}-formatted neutral demographic history of the metapopulation}

\item{No}{(int) effective size of the reference population (ie. number of diploid individuals)}
}
\value{
An array giving the most ancient sweep age (scaled in \eqn{4No}) for local adaptation in each deme.
}
\description{
Automatically determines the most ancient sweep age possible for local adaptation when >=2 populations.
}
\details{
Note that in the situation of a Kingman approximation to the real multicoalescent, we propose to underestimate the most ancient sweep age by \emph{dT(i)} (i.e. the multicoalescent delay approximation, by default \eqn{dT(i) = nSamples(i) - 1}) in order to avoid temporal conflicts with other neutral events specified in the \code{ms}-formatted demographic history.
}
\seealso{
\code{\link{generate_priors}}
}
\keyword{internal}

