% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{get_pls}
\alias{get_pls}
\title{Get model-specific PLS components}
\usage{
get_pls(param, ss, PLS_normalize, removeCollinearCols, PLS_ncomp, plot_PLScv,
  maxPLSncomp = 100, deme, QR_multicollinearity = T)
}
\arguments{
\item{PLS_ncomp}{if NULL, automatically select the fittest number of PLS components, else an integer to set the number of components to a specific value}
}
\keyword{internal}

