% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{getmode}
\alias{getmode}
\title{Calculate the mode of a univariate distribution}
\usage{
getmode(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
A numeric value corresponding to the mode of the interpolated empirical distribution.
}

