% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{gscan}
\alias{gscan}
\title{Ensemble genome scan for sweep detection & sweep age estimation}
\usage{
gscan(X, reftb, firstPos = NULL, lastPos = NULL, minSNP = 10,
  windowSizes = seq(10000, 2e+05, length.out = 20), nSteps = 20,
  discard_extraRange = TRUE)
}
\arguments{
\item{X}{an observed dataset obtained with \code{get_SFS()} or a pseudo-observed dataset obtained with \code{generate_pseudoobs()}}

\item{reftb}{a \code{referenceTable} object, with non-empty \code{PRIORS}, \code{SFS} and \code{DIMREDUC} slots}

\item{firstPos}{(integer) the first position of the genome scan (NULL to automatically start at the first known position)}

\item{lastPos}{(integer) the last position of the genome scan (NULL to automatically stop at the last known position)}

\item{minSNP}{(integer) minimum number of SNPs in the window to consider it valid for inference}

\item{windowSizes}{(array of integers) a vector of window lengths over which genome scans will be performed iteratively, you may consider using the R function seq(minimum_length, maximum_length, length.out = number_of_values) to sample a set of equally-spaced lengths}

\item{nSteps}{(integer) number of overlapping shifts (the higher the number, the more overlapping windows will be introduced, increasing the scan resolution)}

\item{discard_extraRange}{(logical) if TRUE, will discard sweep age estimations going beyond the prior range (TRUE is recommended)}
}
\value{
A list to be further processed with the \code{thin()} function.
}
\seealso{
\code{\link{get_SFS}}, \code{\link{generate_pseudoobs}}, \code{\link{thin}}
}

