% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{linearize}
\alias{linearize}
\title{Merge model-specific multiSFSs into one big matrix}
\usage{
linearize(sfs)
}
\arguments{
\item{sfs}{a list of model-specific multiSFS matrices}
}
\value{
A list of two elements: a vector of model indices; a big matrix containing all merged multiSFSs.
}
\keyword{internal}

