% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{print.observedDataset}
\alias{print.observedDataset}
\title{Print a summary of an \code{observedDataset} object content}
\usage{
print.observedDataset(X)
}
\arguments{
\item{X}{a \code{observedDataset} object (generated with \code{\link{get_SFS}})}
}
\seealso{
\code{\link{get_SFS}}
}

