% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{print.referenceTable}
\alias{print.referenceTable}
\title{Print a summary of a \code{referenceTable} object content}
\usage{
print.referenceTable(X)
}
\arguments{
\item{X}{a \code{referenceTable} object}
}
\seealso{
\code{\link{generate_priors}}
}

