% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{print.validationTable}
\alias{print.validationTable}
\title{Print a summary of a \code{validationTable} object content}
\usage{
print.validationTable(X)
}
\arguments{
\item{X}{a \code{validationTable} object}
}
\seealso{
\code{\link{generate_pseudoobs}}
}

