% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{project_target}
\alias{project_target}
\title{Project an observed multiSFS onto LDA/PLS axes, stored in a referenceTable object}
\usage{
project_target(reftb, target, method, focalIsland = NULL)
}
\arguments{
\item{reftb}{a reference table (class \emph{referenceTable}) with non-NULL DIMREDUC element}

\item{target}{the observed multiSFS}

\item{method}{(string) either "LDA" (model selection) or "PLS" (parameter estimation)}

\item{focalIsland}{index of the focal island}
}
\keyword{internal}

