% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan_core}
\alias{scan_core}
\title{Core function for the genome scan}
\usage{
scan_core(reftb, POS, PAC, wSNP, firstPos, lastPos, minSNP, windowSizes, nSteps,
  infer.age = TRUE, discard_extraRange = TRUE, progressBar = TRUE)
}
\keyword{internal}

