% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{set_paths}
\alias{set_paths}
\title{Set mandatory paths to external softwares}
\usage{
set_paths(tempDir, pyPath, javaPath, msmsPath, pyMSMSPath, pyCLEANMSPath)
}
\arguments{
\item{tempDir}{path to a temporary directory}

\item{pyPath}{path to your \code{python} executable, or the OS environment variable}

\item{javaPath}{path to your \code{java} executable, or the OS environment variable}

\item{msmsPath}{path to the \code{msms} executable, or the OS environment variable}
}

