% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationProcedures.R
\name{sliding_validation}
\alias{sliding_validation}
\title{Perform genome scan on simulated genomic fragments}
\usage{
sliding_validation(valtb, reftb, minSNP = 10, cutoff = 10, tolABC = 0.01,
  windowSlide = NULL, tolGFIT = 0.1, ...)
}
\arguments{
\item{valtb}{a \code{validationTable} object generated by \code{\link{generate_pseudoobs}}}

\item{reftb}{an initialized \code{referenceTable} object with non-empty \code{PRIORS} slot}

\item{minSNP}{(integer) minimum number of within-window SNPs; if the number of within-window SNPs \code{< minSNP}, the function will return \code{NA}}

\item{cutoff}{(numeric) the Bayes Factor cut-off value above which a window is assigned to the fittest model}

\item{tolABC}{(numeric or an array of 2 elements, values between 0 and 1) if a single numeric value, the tolerance ratio for both sweep detection and sweep estimation (see \code{\link{postpr}} and \code{\link{abc}}); if an array of 2 numeric elements, the first one will be the tolerance ratio for sweep detection (see \code{\link{postpr}}) and the second one for sweep estimation (see \code{\link{abc}})}

\item{windowSlide}{(integer) the sliding gap, if \code{NULL} then \emph{windowSlide} will be set equal to the window size (windows will thus slide without overlap)}

\item{tolGFIT}{(numeric between 0 and 1) significance level for the goodness-of-fit test, ie. if \code{p-value(GoF) < tolGFIT}, we reject Ho: "The observed dataset falls within the null distribution of distances"; i.e. the observed dataset is not explainable by a given evolutionary model; if you want to disable the GoF test, set \code{tolGFIT = 0}}

\item{...}{other arguments passed to \code{\link{thin}}, please consult \code{?thin}}
}
\value{
A \code{validationTable} object with an additional \code{ANALYSIS} slot containing the results of the genome scans over the simulated fragments.
}
\description{
This function is equivalent to \code{thin(gscan(...))} except that it works for simulated genomic fragments stored in an \code{validationTable} object.
}
\examples{
Please refer to the vignette.
}

