% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationProcedures.R
\name{sliding_validation_trim}
\alias{sliding_validation_trim}
\title{Validation trimming procedure for sweep age point estimation}
\usage{
sliding_validation_trim(valtb, reftb, deme, method, which = NULL, ...)
}

