% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationProcedures.R
\name{sliding_validation_trim_MODIF}
\alias{sliding_validation_trim_MODIF}
\title{Validation trimming procedure for sweep age point estimation}
\usage{
sliding_validation_trim_MODIF(valtb, which = NULL, method, bwadjust = NULL,
  discardSingleWindow = FALSE, POD = NULL, minWindows = 1)
}

