% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationProcedures.R
\name{sliding_validation_trim_REF}
\alias{sliding_validation_trim_REF}
\title{Validation trimming procedure for sweep age point estimation}
\usage{
sliding_validation_trim_REF(valtb, method, bwadjust = NULL,
  discardSingleWindow = FALSE)
}

