% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{summary.observedDataset}
\alias{summary.observedDataset}
\title{Genomic SNP density in the observed dataset}
\usage{
summary.observedDataset(obs, windowSize)
}
\arguments{
\item{obs}{an \code{observedDataset} object, generated by \code{\link{get_SFS}}}

\item{windowSize}{(integer) a window size (in base pairs)}
}
\value{
A summary of the distribution of within-window SNP counts.
}
\description{
Given a window length, compute the distribution of within-window SNP counts in the observed dataset. This function is useful to get an idea about the SNP density in the observed dataset to fix (i) an appropriate window size and (ii) the minimum number of SNPs to retain a window for downstream analyses.
}
\examples{
load(system.file("data", "UNFOLDED_POLARIZED_2dSFS_CEU-LWK_20-20_chr2", package = "McSwan"))
summary(obs, windowSize = 10000)
}
\seealso{
\code{\link{generate_priors}}
}

