% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{summary.validationResult}
\alias{summary.validationResult}
\title{Summary of the sliding validation}
\usage{
summary.validationResult(X, valtb, file)
}
\arguments{
\item{X}{a \code{validationResult} object (i.e. returned by \code{\link{gscan}})}

\item{valtb}{a \code{validationTable} object (generated by \code{generate_pseudoobs}, the same object used in the previous \code{\link{gscan}} call)}

\item{file}{path to a file where validation graphics will be plotted}
}
\value{
A list of 6 elements: \itemize{ 
\item \code{absolute.confusion.matrix}: confusion matrix with absolute counts
\item \code{absolute.relative.matrix}: confusion matrix with relative counts
\item \code{performance.rates}: specificity (TNR), sensitivity (TPR) and false discovery rate (FDR)
\item \code{number.of.sweeps}: number of sweep contigs detected in each POD (ideally, should be 0 for the neutral model and 1 for the selective models)
\item \code{sweep.inclusiveness}: number of times the detected sweeps encompass the beneficial mutation
\item \code{NRMSE}: normalized/standardized root mean square error
}
}
\examples{
Please refer to the vignette.
}

