% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationProcedures.R
\name{summary_cv}
\alias{summary_cv}
\title{Detect sweeps and estimate sweep age}
\usage{
summary_cv(X, file)
}
\arguments{
\item{X}{either the output from simulate_obs || a VCF filepath || a multiSFS}
}

