% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thin_internal}
\alias{thin_internal}
\title{Thinning - internal procedure}
\usage{
thin_internal(x, reftb, method = "contig", stat = "mean", bwadjust = 1,
  minWindows = 1, plot_thinning = F, epsilon = 1e-15,
  summary_stat = "mean")
}
\keyword{internal}

