% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thin_SIMPLEGOOD}
\alias{thin_SIMPLEGOOD}
\title{Thinning}
\usage{
thin_SIMPLEGOOD(x, reftb, method = "contig", stat = "mean", bwadjust = 1,
  minWindows = 1, plot_thinning = F, epsilon = 1e-15,
  summary_stat = "mean")
}

