% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{analyze}
\alias{analyze}
\title{Sweep detection & sweep age estimation for a single genomic region}
\usage{
analyze(target, reftb, minSNP = 2, tolABC = 0.01, tolGFIT = 0.05,
  cutoff = 5, plot_simCloud = FALSE, verbose = F)
}
\arguments{
\item{target}{multidimensional joint site-frequency-spectra of the target genomic window}

\item{reftb}{a \code{referenceTable} object, with non-empty \code{PRIORS}, \code{SFS} and \code{DIMREDUC} elements}

\item{minSNP}{(integer) minimum number of within-window SNPs; if the number of within-window SNPs \code{< minSNP}, the function will return \code{NA}}

\item{tolABC}{(numeric or an array of 2 elements, values between 0 and 1) if a single numeric value, the tolerance ratio for both sweep detection and sweep estimation (see \code{\link{postpr}} and \code{\link{abc}}); if an array of 2 numeric elements, the first one will be the tolerance ratio for sweep detection (see \code{\link{postpr}}) and the second one for sweep estimation (see \code{\link{abc}})}

\item{tolGFIT}{(numeric between 0 and 1) significance level for the goodness-of-fit test, ie. if \code{p-value(GoF) < tolGFIT}, we reject Ho: "The observed dataset falls within the null distribution of distances"; i.e. the observed dataset is not explainable by a given evolutionary model; if you want to disable the GoF test, set \code{tolGFIT = 0}}

\item{cutoff}{(numeric) the Bayes Factor cut-off value above which a window is assigned to the fittest model}

\item{plot_simCloud}{(logical) plot the observed dataset amid the simulated ones, in the space of the retained LDA components}

\item{verbose}{(logical) verbose mode}
}
\value{
A list of 4 elements:
\itemize{
\item \code{sweepingIsland} the population which has experienced a recent positive sweep
\item \code{BayesFactor} the Bayes Factor in favour of the model
\item \code{params} point estimate for (\code{sweepAge})
\item \code{ic} a vector giving the lower and upper limits of the 95\% credible intervals for (\code{sweepAge}
}
}
\seealso{
\code{\link{gscan}} for iterating this function by sliding windows along the genome
}
\keyword{internal}

