% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{Performs multicoalescence SIMPLE PARTITION WITHOUT MIGRATION, given a set of parameter priors}
\usage{
coalesce(x, method = "partition", nRep = 1, SAA = 1, execute = TRUE,
  phyclust = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{an initialized reference table (class \emph{referenceTable}).}

\item{nRep}{(optional) number of repetitions to approximate the sweep (default 100)}
}
\value{
A reference table with all simulated site frequency spectra.
}
\description{
NA
}
\details{
Past resizings are always set in reference to  \emph{No}.
}

