% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\title{Coalescent simulations , given a set of parameter priors}
\usage{
coalesce(x, execute = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{an initialized \code{referenceTable} object}

\item{execute}{(logical) whether to execute the simulations}

\item{verbose}{(logical) verbose mode}
}
\value{
A reference table with all simulated site frequency spectra stored in the \code{SFS} slot.
}
\description{
Given a \code{referenceTable} object containing prior parameter values and a neutral demographic history, this function generates the command lines of neutral and selective models and execute them with Hudson's \emph{MS} coalescent simulator.
}
\details{
Past population size rescalings are always set in reference to \code{No}. If you want to see the demographic command lines for each model before simulating, set \code{execute = FALSE} and \code{verbose = TRUE}.
}
\examples{
Please refer to the vignette.
}
\references{
Hudson, R.R. (2002). Generating samples under a Wright-Fisher neutral model of genetic variation. Bioinformatics, 18, 337-338.
}
\seealso{
\code{\link{generate_priors}}, \code{\link{dim_reduction}}
}

