% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceTable_manipulations.R
\name{combine.referenceTable}
\alias{combine.referenceTable}
\title{Combine referenceTable objects}
\usage{
\method{combine}{referenceTable}(V, v)
}
\arguments{
\item{V}{a first \code{referenceTable} object with empty \code{DIMREDUC} slot}

\item{v}{a second \code{referenceTable} object with empty \code{DIMREDUC} slot}
}
\value{
A \code{referenceTable} object combining the two original \code{referenceTable} objects.
}
\description{
Combine two \code{referenceTable} objects prior to machine learning (i.e. prior to the \code{\link{dim_reduction}} call).
}
\seealso{
\code{\link{generate_priors}}, \code{\link{coalesce}}, \code{\link{dim_reduction}}
}

