% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_manipulations.R
\name{convert_VCF}
\alias{convert_VCF}
\title{convert a VCF to a proper format}
\usage{
convert_VCF(vcfPath, pops, reftb, outPath, chromosome, minQUAL = NULL)
}
\arguments{
\item{pops}{a dataframe containing two columns: \emph{sample.ID} & \emph{population.ID}. Note that \emph{population.ID} needs to be an integer and that \emph{population.ID} must correspond to the index of the island specified in your \emph{ms} command}
}

