% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{dim_reduction}
\alias{dim_reduction}
\title{Dimension reduction prior to approximate Bayesian computation}
\usage{
dim_reduction(x, removeCollinearCols = TRUE, compress_SFS = TRUE,
  LDA_xPC = 0.1, LDA_sizePerModel = NULL, PLS_normalize = TRUE,
  PLS_ncomp = NULL, PLS_ncomp_method = "elbow", PLS_maxncomp = 100,
  PLS_plotCV = TRUE)
}
\arguments{
\item{x}{a \code{referenceTable} object with non-empty \code{PRIORS} and \code{SFS} slots}

\item{removeCollinearCols}{(logical) whether to remove collinear columns of the site frequency spectra prior to LDA & PLS dimension reduction (TRUE is recommended)}

\item{LDA_xPC}{(float between 0 and 1) number of principal components to retain (as a fraction of the number of cells in the SFS) to do subsequent LDA}

\item{LDA_sizePerModel}{(integer) number of simulations per model to consider for training the PCA-LDA; \emph{NULL} (by default) forces to use all data}

\item{PLS_normalize}{(logical) whether to normalize the spectra prior to PLS analysis (TRUE is recommended)}

\item{PLS_ncomp}{(integer) if NULL, automatically select the fittest number of PLS components, else an integer to manually set the number of retained components}

\item{PLS_ncomp_method}{(string) method to automatically select the best number of features: either "\emph{elbow}", "\emph{onesigma}" or "\emph{randomization}" (cf. \code{selectNcomp} in package \code{pls})}

\item{PLS_maxncomp}{(integer) maximum number of PLS components to calculate before selecting the best set of features}

\item{plot_PLScv}{(logical) if TRUE, will render the cross-validation graphs of the PLS}
}
\value{
An object of class \code{referenceTable} with filled-in \code{DIMREDUC} slot.
}
\description{
Reduces the dimension of the summary statistics (joint multidimensional allele frequency spectra) prior to ABC model selection (through linear discriminant analysis) and ABC parameter estimation (through model-specific partial least square regressions). This function is to be run after \code{\link{coalesce}}.
}
\examples{
Please refer to the vignette.
}
\seealso{
\code{\link{coalesce}}, \code{\link{plsr}}, \code{\link{lda}}, \code{\link{abc}}
}

