% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_priors.R
\name{generate_priors}
\alias{generate_priors}
\title{Initialize the reference table}
\usage{
generate_priors(msDemography, No, fold, windowSize = 1e+05, nSimul = 10000,
  restrictDemesTo = NULL, sweepAge = NULL)
}
\arguments{
\item{msDemography}{(string) the neutral demographic history of the metapopulation, formatted according to Hudson's \emph{MS} conventions, where \eqn{\theta = 4N_o\mu} with \eqn{N_o} the diploid effective size; times are scaled in units of \eqn{4N_o} and population size rescaling in units of \eqn{N_o}; note that the demographic history can involve unsampled (ghost) populations}

\item{No}{(integer) the \bold{diploid} effective size of the reference population}

\item{fold}{(logical) whether allele counts are to be defined in reference to the \bold{minor} alleles (\code{fold = TRUE}, producing \bold{folded} site frequency spectra) or to the \bold{derived} alleles (\code{fold = FALSE}), producing \bold{unfolded} site frequency spectra. In the derived allele case, you must ensure that the SNPs in your VCF dataset are polarized relatively to some ancestral individuals (thus genotypes are defined in terms of ANCESTRAL/DERIVED instead of REF/ALT).}

\item{windowSize}{(integer) length of the sequences (in bp) to simulate, the larger the better (> 100.000)}

\item{nSimul}{(integer) number of simulations to perform for each model (the larger, the better but beware memory usage! previous tests suggested that 20.000 were sufficient for moderately complex demographic scenarios)}

\item{restrictDemesTo}{(array of integers) by default, McSwan will detect population-specific sweeps across all the populations specified in the \code{MS} command; however, if you want to restrict the analysis to a particular subset of demes, provide their \bold{indices} in a vector; note that population indices must correspond to their position under the \code{-I} switch of the \code{MS} command, and note that the index of the first population is \bold{1}}

\item{sweepAge}{(special list) prior distribution for the sweep ages (scaled in generations before present); if \code{NULL} and the number of populations is superior to 2, the prior distribution ranges will be automatically determined, otherwise it is mandatory to specify the distribution manually, see \code{Details})}
}
\value{
An object of class \emph{referenceTable} with initialized \code{PRIOR} slot.
}
\description{
Semi-automatic initialization of the prior distributions for all model parameters.
}
\details{
Prior distributions must be specified using the following syntax: \code{list("P", arg1, arg2)} with \emph{P} the name of the distribution function (e.g. \code{\link{runif}} for the uniform distribution, \code{\link{rlogunif}} for the log-uniform; please make sure you have quoted the function name and removed the argument brackets); \emph{arg1} and \emph{arg2} respectively the first and second arguments of the function (e.g. for \code{runif} will be the lower and upper limits of the distribution). Note that if the upper and lower limits of the distribution are distant by more than 3 units in the \emph{log10} scale, it is recommended to use the \code{rlogunif} distribution.

If you manually provide the prior distributions for the \bold{sweep ages}, you have two options: \itemize{
\item either specify a single \code{list("P", arg1, arg2)} which will be the common distribution set \bold{for all} the demes,
\item or specify a list of distribution-lists, each distribution-list corresponding to the sweep age distribution for a specific deme (indexed as they appear in the \code{ms} command); for instance, for 2 demes, one would specify: \code{list(list("rlogunif", T_1, TT_1), list("runif", T_2, TT_2))}. Please note that you must specify distributions for every deme, even if you have restricted the analysis to specific demes \emph{via} the \code{restrictDemesTo} option.
}
}
\examples{
Please refer to the vignette.
}
\references{
Hudson, R.R. (2007) \code{ms} - a program for generating samples under neutral models \url{http://home.uchicago.edu/rhudson1/source/mksamples/msdir/msdoc.pdf}
}
\seealso{
\code{\link{coalesce}}
}

