% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{get_dists}
\alias{get_dists}
\title{Get the null distribution of distances within a model}
\usage{
get_dists(x, n_dist_bstp_per_model, abcTolerance)
}
\arguments{
\item{x}{a model-specific multiSFS matrix}

\item{n_dist_bstp_per_model}{number of distances to calculate per model to build the null distribution}
}
\value{
A vector of distances, giving an empirical null distribution.
}
\keyword{internal}

