% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{get_events}
\alias{get_events}
\title{Extract all demographic events before an event time-scaled age and involving some specific island}
\usage{
get_events(ms, islandIndex, eventAge)
}
\arguments{
\item{ms}{the ms-formatted demographic history}

\item{islandIndex}{the index of the focal island}

\item{eventAge}{the age of the event, scaled in 4No}
}
\value{
a string of ms-formatted history events affecting the twin subisland
#@export
}
\description{
does not handle -es -eM -eG -eN -N -M -ema
}
\keyword{internal}

