% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{get_size}
\alias{get_size}
\title{Extract the deme effective sizes just before an event time-scaled age}
\usage{
get_size(ms, islandIndex, eventAge)
}
\arguments{
\item{ms}{the ms-formatted demographic history}

\item{islandIndex}{the index of the focal island}

\item{eventAge}{the age of the event, scaled in units of 4No}
}
\value{
the resizing coefficient, in reference to No
}
\keyword{internal}

