% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{gscan}
\alias{gscan}
\title{Performs a genome scan, by detecting sweeps along sliding windows and estimating sweep ages}
\usage{
gscan(X, reftb, minSNP, startPos = NULL, lastPos = NULL,
  windowSlide = NULL, printProgress = T, ...)
}
\arguments{
\item{X}{either an object with class \emph{observedDataset} produced by \code{get_SFS()} or with class \emph{validationTable}}

\item{reftb}{a \emph{referenceTable} object, with activated \code{PRIORS}, \code{SFS} and \code{DIMREDUC} elements}

\item{startPos}{(integer) the starting position of the genome scan}

\item{lastPos}{(integer) the ending position of the genome scan}

\item{windowSlide}{(integer) the sliding gap, if \code{NULL} then \emph{windowSlide} will be automatically set as equal the window size (therefore windows will not overlap)}

\item{...}{other arguments passed to \code{\link{analyze}}}
}
\value{
A dataframe with each row corresponding to one window and with the following columns: \itemize{
\item \code{start.pos} the starting position of the window
\item \code{end.pos} the ending position of the window
\item \code{deme.under.sel} the deme which has been detected to have undergone recent positive sweep 
\item \code{BayesFactor} the Bayes Factor for the sweep model of the sweeping deme (ie. \code{deme.under.sel})
\item \code{sweepAge} the point estimate of the sweep age
\item \code{recRate} the point estimate of the recombination rate between adjacent bases
\item \code{sweepAge.IC.low} the lower bound of the 95\% credible interval for the sweep age estimation
\item \code{sweepAge.IC.up} the upper bound of the 95\% credible interval for the sweep age estimation
\item \code{recRate.IC.low} the lower bound of the 95\% credible interval for the recombination rate estimation
\item \code{recRate.IC.up} the upper bound of the 95\% credible interval for the recombination rate estimation
\item \code{failure} if the sweep detection has failed, specifies the reason: either because the window failed the requirement for a minimum number of SNPs (\code{minSNP}); failed the goodness-of-fit test (\code{goodnessOfFit}); failed the ABC estimation because of zero variance in the LDA/PLS components for instance (\code{ABC}); failed the discriminability between equally likely models (\code{modelDiscrimination}).
}
}
\description{
This function performs a genome scan along sliding windows, for each window it estimates the Bayes Factor of the most likely sweep model in a specific deme \emph{vs.} the neutral model. If a sweep is detected, it carries on estimating the age of the sweep.
}
\seealso{
\code{\link{thin}}, \code{\link{get_SFS}}
}

