% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{impute}
\alias{impute}
\title{impute model based on Bayes Factors}
\usage{
impute(postproba, cutoff)
}
\details{
postproba: a vector of model posterior probabilities, the first element must be the neutral model

this vector must be NAMED by the model names ("i0", "i1", "i2", ... by default)
}
\keyword{internal}

