% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{print.observedDataset}
\alias{print.observedDataset}
\title{Print a summary of a \code{observedDataset} object content}
\usage{
\method{print}{observedDataset}(X)
}
\arguments{
\item{X}{a \code{observedDataset} object}
}
\seealso{
\code{\link{get_SFS}}
}

