% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellenaous.R
\name{set_session}
\alias{set_session}
\title{Set mandatory paths to external softwares}
\usage{
set_session(tempDir, javaPath)
}
\arguments{
\item{tempDir}{path to a temporary directory}

\item{javaPath}{path to your \code{java} executable, or the OS environment variable}

\item{pyPath}{path to your \code{python} executable, or the OS environment variable}

\item{msmsPath}{path to the \code{msms} executable, or the OS environment variable}
}

