% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationProcedures.R
\name{sliding_validation}
\alias{sliding_validation}
\title{Validation procedure on pseudo-chromosomes}
\usage{
sliding_validation(valtb, reftb, method, stat = "mean", minSNP = 10,
  cutoff = 10, tolABC = 0.01, windowSlide = NULL, tolGFIT = 0.1,
  plot_simCloud = F, plot_thinning = F, bwadjust = 1, minWindows = 1,
  summary_stat = "mean", ...)
}

