% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{summary.validationResult}
\alias{summary.validationResult}
\title{Summary of the sliding validation}
\usage{
summary.validationResult(X, valtb, SNP_wise_estimates = TRUE, file)
}
\arguments{
\item{X}{a \code{validationTable} object with non-empty \code{ANALYSIS} slot (ie. obtained after executing the \code{sliding_validation} function)}

\item{file}{the name of a file where validation graphics are plotted}
}
\value{
A list of 6 elements: \itemize{ 
\item \code{absolute.confusion.matrix} confusion matrix with absolute counts
\item \code{absolute.relative.matrix} confusion matrix with absolute counts
\item \code{performance rates} specificity (TNR), sensitivity (TPR) and false discovery rate (FDR)
\item \code{number.of.sweeps} number of contiguous sweeps detected in each simulated genomic fragment
\item \code{sweep.inclusiveness} number of times the detected sweeps encompass the beneficial mutation
\item \code{NRMSE} normalized/standardized root mean square error
}
}
\examples{
Please refer to the vignette.
}
\seealso{
\code{\link{sliding_validation}}
}

