% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thin}
\alias{thin}
\title{Merge contiguous sweeps}
\usage{
thin(scanResult, reftb, X, max_L = 1e+06, signif.threshold = 0.5,
  maxIter = 1000, stat = mean)
}
\arguments{
\item{reftb}{an initialized \code{referenceTable} object}

\item{x}{a dataframe returned by \code{\link{gscan}}}

\item{bwadjust}{(numeric) bandwidth coefficient, the algorithm will extend contiguity as far as \code{bwadjust * windowSize} from any focal window}

\item{minWindows}{(integer) the minimum number of contiguous windows required to perform contiguity search}

\item{summary_stat}{(string) the central statistic to average the parameter estimates of merged contiguous windows ("mean" recommended)}

\item{plot_thinning}{(logical) plot a representation of the genomic fragment with the contiguous sweep regions}
}
\value{
Returns a dataframe of the merged contiguous sweeps. One sweep per line, and in columns:
\itemize{ 
\item \code{\bold{sweep.center}} center of the contiguous sweep region
\item \code{\bold{sweep.lbound}} first position of the contiguous sweep region
\item \code{\bold{sweep.rbound}} last position of the contiguous sweep region
\item \code{\bold{BF}} aggregated Bayes Factor of the contiguous sweep region
\item \code{\bold{deme}} population detected to have experienced a selective sweep (given as "i" & index of the population in the \emph{MS} command)
\item \code{\bold{sweepAge}} posterior estimate for the sweep age
\item \code{\bold{sweepAge.IC.low}} lower boundary of the 95% credible interval for the sweep age estimation
\item \code{\bold{sweepAge.IC.up}} upper boundary of the 95% credible interval for the sweep age estimation
}
}
\description{
This function merges contiguous regions detected by \code{\link{gscan}} as having experienced a selective sweep.
}
\examples{
Please refer to the vignette.
}
\seealso{
\code{\link{gscan}}
}

